+++ Lib/UI/Dom.js
@@ -82,6 +82,26 @@ export class element {
 		return this._elem.type;
 	}
 
+	// sets the src attribute of the element
+	src(src) {
+		this._elem.src = src;
+		return this;
+	}
+	// gets the src attribute of the element
+	get_src() {
+		return this._elem.src;
+	}
+
+	// sets the background image of the element
+	background_image(src) {
+		this._elem.style.backgroundImage = `url(${src})`;
+		return this;
+	}
+	// gets the background image of the element
+	get_background_image() {
+		return this._elem.style.backgroundImage;
+	}
+
 	// sets the given attribute of the element
 	attr(name, value) {
 		this._elem.setAttribute(name, value);