+++ js/constants.js
@@ -1,0 +1,6 @@
+export const SELL = 0;
+export const BUY = 1;
+
+export const LOGIN = 0;
+export const SUCCESS = 1;
+export const PRODUCTS = 2;